fix register user

+++ Source/ISA/ISA/Controllers/AccountController.cs
@@ -131,8 +131,11 @@ namespace ISA.Controllers
                 var user = new ApplicationUser { UserName = model.Email, Email = model.Email };
 
                 // ONLY FOR TESTING PURPOSE!!! MADE BY LUKA
-                var userToDelete = _userManager.Users.First();
-                await _userManager.DeleteAsync(userToDelete);
+                var userToDelete = _userManager.Users.FirstOrDefault();
+                if (!(userToDelete is null))
+                {
+                    await _userManager.DeleteAsync(userToDelete);
+                }
 
                 var result = await _userManager.CreateAsync(user, model.Password);
                 if (result.Succeeded)
